[Script] Casting int to short
- Fixed warning

[ci skip]

+++ src/Components/Modules/Client.cpp
@@ -192,7 +192,7 @@ namespace Components
 
 			if (client->state >= 3)
 			{
-				client->ping = ping;
+				client->ping = (short)ping;
 			}
 		});
 	}
